+++ pydantic_aiohttp/__init__.py
@@ -1,6 +1,7 @@
-__version__ = '0.6.2'
+__version__ = '0.7.0'
 
 from . import encoders
 from . import utils
 from .client import Client
 from .errors import *
+from .types import *